Improve default configuration and remove passwords
I thought it would be useful to specify the password for a local server
setup, but that is a bit confusing because it diverges from the cloud
configuration setup and also looks like a real password etc. bad
behavior.

It is necessary to keep the two configuration separate since there are
tests for both the cloud and local server which would both execute at
the same time. With `LOAD_ENV=1 ...`, the same values would be used
for both, which would cause some tests to fail with unexpected requests.